% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetCalibrate.R
\name{predict.homCalibrate}
\alias{predict.homCalibrate}
\title{Predictions using a homoskedastic model object (of class \code{homCalibrate})}
\usage{
\method{predict}{homCalibrate}(object, x, xprime = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{homCalibrate}; e.g., as returned by \code{\link[HetCalibrate]{mleHomCalibrate}}}

\item{x}{matrix of designs locations to predict at (one point per row)}

\item{xprime}{optional second matrix of predictive locations to obtain the predictive covariance matrix between \code{x} and \code{xprime}}

\item{...}{no other argument for this method}
}
\value{
list with elements
\itemize{
\item \code{mean}: kriging mean;
\item \code{sd2}: kriging variance (filtered, e.g. without the nugget value)
\item \code{cov}: predictive covariance matrix between \code{x} and \code{xprime}
\item \code{nugs}: nugget value at each prediction location, for consistency with \code{\link[HetCalibrate]{mleHomCalibrate}}.
}
}
\description{
Predictions using a homoskedastic model object (of class \code{homCalibrate})
}
\details{
The full predictive variance corresponds to the sum of \code{sd2} and \code{nugs}. See \code{\link[HetCalibrate]{mleHomCalibrate}} for examples. This code is modified from the source code of \code{\link[hetGP]{predict.homGP}}.
}
